backend/requirements.txt
asgiref==3.6.0
astroid==2.15.5
dill==0.3.6
Django==4.2.1
django-cors-headers==4.0.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
Pillow==9.5.0
platformdirs==3.5.1
pylint==2.17.4
pytz==2023.3
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.11.8
typing_extensions==4.6.1
wrapt==1.15.0
